requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
paramiko==3.4.0
psutil==5.9.6
APScheduler==3.10.4
gunicorn==21.2.0
